pyproject.toml
[project]
name = "chess-gpt"
version = "0.1.0"
description = "Training Llama3 to play chess."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.2.1",
    "datasets>=3.2.0",
    "deepspeed>=0.16.1",
    "evaluate>=0.4.3",
    "ruff>=0.8.3",
    "setuptools>=75.6.0",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "transformers>=4.47.0",
    "wandb>=0.19.1",
    "scikit-learn>=1.0.1",
    "liger-kernel>=0.5.2",
    "flash-attn>=2.7.2.post1",
    "debugpy>=1.8.11",
    "python-chess>=1.999",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu124", marker = "platform_system != 'Darwin'"},
]
torchvision = [
  { index = "pytorch-cu124", marker = "platform_system != 'Darwin'"},
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
